Add back the modernized example.
This upgrades my initial demo to make use of react 16 and rxjs 6.
I've refactored the fibonacci example into it's own component,
so it can be displayed multiple times.
With this, I can display a synchronized examples with, and without cache.

I've also made the WorkerService available as react hook,
which makes it way easier to use. All the rxjs subscription logic is now
contained within the hook and doesn't leak into the component.